Rest restructure rest api test, service catalogs
